Move everything into bot.py
