+++ Dockerfile
@@ -7,4 +7,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 COPY . .
 
-CMD [ "python", "-u", "./src/server.py" ]
+CMD [ "python", "-u", "./src/bot.py" ]
